Fix target count on Infant child and pregnancy visits (#43)
From v3.8, targets are by default deduped by contact. If more than one
report applies and they are within the same period, they'll be counted
as 1 since the emission id is based on the contact. To change this, we
need to set the idType to report so that all applicable reports count.

Co-authored-by: Derick Lung'aho <derick.lungaho@users.noreply.github.com>